# schemas/openapi/public_v1.yaml
openapi: 3.1.0
info:
  title: RoadPulse Public API
  description: RoadPulse is a flood-aware mobility intelligence layer for Vietnam.
    This OpenAPI document covers the public ``/v1/*`` surface — three-route planning,
    batch ETA, isochrones, flood risk, site selection, fleet matching and the parametric
    insurance trigger feed. See ``docs/api/`` for SDK snippets.
  version: 0.1.0
paths:
...
  /v1/route:
    post:
      tags:
      - route
      summary: Plan a flood-aware route with fast/safe/eco variants
      operationId: post_route_v1_route_post
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/flood-risk:
    get:
      tags:
      - flood
      summary: Return current/forecast flood scores for a set of hexes or bbox
      operationId: get_flood_risk_v1_flood_risk_get
      parameters:
      - name: hex_ids
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          title: Hex Ids
      - name: bbox
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            prefixItems:
            - type: number
            - type: number
            - type: number
            - type: number
            minItems: 4
            maxItems: 4
          - type: 'null'
          description: minLng,minLat,maxLng,maxLat
          title: Bbox
        description: minLng,minLat,maxLng,maxLat
      - name: horizon
        in: query
        required: false
        schema:
          enum:
          - now
          - 1h
          - 3h
          - 6h
          type: string
          default: now
          title: Horizon
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FloodRiskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EtaConfidence:
      type: string
      enum:
      - low
      - medium
      - high
      title: EtaConfidence
      description: Categorical confidence label attached to ETA predictions.
...
    FloodOverlayPoint:
      properties:
        hex_id:
          type: string
          title: Hex Id
        centroid:
          $ref: '#/components/schemas/LatLon'
        score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Score
        horizon:
          type: string
          enum:
          - now
          - 1h
          - 3h
          - 6h
          title: Horizon
          default: now
      type: object
      required:
      - hex_id
      - centroid
      - score
      title: FloodOverlayPoint
    FloodRiskResponse:
      properties:
        horizon:
          type: string
          enum:
          - now
          - 1h
          - 3h
          - 6h
          title: Horizon
        generated_at:
          type: string
          format: date-time
          title: Generated At
        hexes:
          items:
            $ref: '#/components/schemas/FloodOverlayPoint'
          type: array
          title: Hexes
      type: object
      required:
      - horizon
      - generated_at
      - hexes
      title: FloodRiskResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LatLon:
      properties:
        lat:
          type: number
          maximum: 90.0
          minimum: -90.0
          title: Lat
        lng:
          type: number
          maximum: 180.0
          minimum: -180.0
          title: Lng
      additionalProperties: false
      type: object
      required:
      - lat
      - lng
      title: LatLon
      description: A geographic coordinate (WGS-84). Latitude first, longitude second.
    RouteMode:
      type: string
      enum:
      - motorbike
      - car
      - truck
      - bicycle
      title: RouteMode
      description: Mode of transport supported by the routing engine.
    RouteRequest:
      properties:
        origin:
          $ref: '#/components/schemas/LatLon'
        destination:
          $ref: '#/components/schemas/LatLon'
        mode:
          $ref: '#/components/schemas/RouteMode'
          default: motorbike
        depart_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Depart At
        time_of_day:
          $ref: '#/components/schemas/TimeOfDay'
          default: now
        avoid_tolls:
          type: boolean
          title: Avoid Tolls
          default: false
        avoid_flood:
          type: boolean
          title: Avoid Flood
          default: true
        locale:
          type: string
          title: Locale
          default: vi
        ride_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Ride Id
      additionalProperties: false
      type: object
      required:
      - origin
      - destination
      title: RouteRequest
    RouteResponse:
      properties:
        request_id:
          type: string
          title: Request Id
        generated_at:
          type: string
          format: date-time
          title: Generated At
        variants:
          items:
            $ref: '#/components/schemas/RouteVariant'
          type: array
          title: Variants
        flood_overlay:
          items:
            $ref: '#/components/schemas/FloodOverlayPoint'
          type: array
          title: Flood Overlay
        weather_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Weather Note
      type: object
      required:
      - request_id
      - generated_at
      - variants
      - flood_overlay
      title: RouteResponse
    RouteStep:
      properties:
        instruction:
          type: string
          title: Instruction
        distance_m:
          type: number
          title: Distance M
        duration_s:
          type: number
          title: Duration S
        bearing_deg:
          type: number
          title: Bearing Deg
        geometry:
          items:
            $ref: '#/components/schemas/LatLon'
          type: array
          title: Geometry
      type: object
      required:
      - instruction
      - distance_m
      - duration_s
      - bearing_deg
      - geometry
      title: RouteStep
    RouteVariant:
      properties:
        name:
          type: string
          enum:
          - fast
          - safe
          - eco
          title: Name
        distance_m:
          type: number
          title: Distance M
        duration_s:
          type: number
          title: Duration S
        free_flow_s:
          type: number
          title: Free Flow S
        flood_score:
          type: number
          title: Flood Score
        eco_score:
          type: number
          title: Eco Score
        toll_vnd:
          type: integer
          title: Toll Vnd
        co2_g:
          type: number
          title: Co2 G
        eta_p10_s:
          type: number
          title: Eta P10 S
        eta_p90_s:
          type: number
          title: Eta P90 S
        eta_confidence:
          $ref: '#/components/schemas/EtaConfidence'
        geometry:
          items:
            $ref: '#/components/schemas/LatLon'
          type: array
          title: Geometry
        steps:
          items:
            $ref: '#/components/schemas/RouteStep'
          type: array
          title: Steps
        hex_path:
          items:
            type: string
          type: array
          title: Hex Path
        notes:
          items:
            type: string
          type: array
          title: Notes
      type: object
      required:
      - name
      - distance_m
      - duration_s
      - free_flow_s
      - flood_score
      - eco_score
      - toll_vnd
      - co2_g
      - eta_p10_s
      - eta_p90_s
      - eta_confidence
      - geometry
      - steps
      - hex_path
      title: RouteVariant
...
    TimeOfDay:
      type: string
      enum:
      - now
      - morning
      - noon
      - evening
      - night
      title: TimeOfDay
      description: Departure time bucket. ``NOW`` means the request is for an immediate
        trip.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError